refactor: new folder for enums
