+++ src/main/java/ru/yandex/practicum/filmorate/model/Event.java
@@ -2,6 +2,8 @@ package ru.yandex.practicum.filmorate.model;
 
 import lombok.*;
 import lombok.experimental.FieldDefaults;
+import ru.yandex.practicum.filmorate.model.enums.EventType;
+import ru.yandex.practicum.filmorate.model.enums.Operation;
 
 import javax.validation.constraints.NotEmpty;
 import java.io.Serializable;
