--- src/main/java/ru/yandex/practicum/filmorate/model/EventType.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.model;
-
-public enum EventType {
-    LIKE,
-    REVIEW,
-    FRIEND
-}
