--- src/main/java/ru/yandex/practicum/filmorate/model/Operation.java
@@ -1,7 +1,0 @@
-package ru.yandex.practicum.filmorate.model;
-
-public enum Operation {
-    REMOVE,
-    ADD,
-    UPDATE
-}
